requirements.txt
	CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==39.0.0
cycler==0.10.0
decorator==5.1.1
deepl==1.13.0
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
firebase==3.0.1
firebase-admin==6.1.0
Flask==2.2.2
Flask-DotEnv==0.1.2
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-cloud-vision==3.3.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h5py==3.8.0
httplib2==0.21.0
idna==2.10
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
networkx==3.0
ninja==1.11.1
numpy==1.24.1
opencv-python==4.7.0.68
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
PyJWT==2.6.0
pyparsing==2.4.7
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
roboflow==0.2.29
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
tifffile==2023.2.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.2
wget==3.2
zipp==3.12.0
